requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
deepface
tf-keras
